So I got the Ai shooting at me, but...
It seems that the overlap of the magic projectile is hitting the instigating pawn. I used GetInstigator in base class and verified who the instigator is, but it's showing up as nullptr basically. meaning despite the AI shooting the projectile, there is no assigned instigator. Sooo...not sure wtf to do with that information since it's not overlapping and reacting to the player when using (OtherActor && OtherActor != GetInstigator())