Create a gamemode override, add Attribute to AI. Add DifficultyCurve...
Created a difficulty Curve to increase maximum number of spawned bots over a period of time. 

Created a custom GameMode class and BP to implement and override the default autogenerated GameMode. 

Modified the MagicProjectile and TargetDummy_BP. 
Made Adjustments to the damage popup widget that ultimately had no effect on it not working lol. 
Made changes to the TestLevel. 

Created an EQS Query to find a spawn point for bots.

Created a Query context to reference all players. 

Created a Query context to reference target actor. 

Generated a Testing Pawn which allows you to see EQS Query locations and results.